# PERF/COMPAT: define platform int to np.intp