--- a/pandas/indexes/base.py
+++ b/pandas/indexes/base.py
@@ -2859,6 +2859,8 @@ class Index(IndexOpsMixin, StringAccessorMixin, PandasObject):
             left_indexer, right_indexer = right_indexer, left_indexer
 
         if return_indexers:
+            left_indexer = None if left_indexer is None else _ensure_platform_int(left_indexer)
+            right_indexer = None if right_indexer is None else _ensure_platform_int(right_indexer)
             return join_index, left_indexer, right_indexer
         else:
             return join_index
@@ -2902,6 +2904,8 @@ class Index(IndexOpsMixin, StringAccessorMixin, PandasObject):
             join_index = self._wrap_joined_index(join_index, other)
 
         if return_indexers:
+            lidx = None if lidx is None else _ensure_platform_int(lidx)
+            ridx = None if ridx is None else _ensure_platform_int(ridx)
             return join_index, lidx, ridx
         else:
             return join_index