--- a/pandas/tests/indexes/common.py
+++ b/pandas/tests/indexes/common.py
@@ -110,7 +110,7 @@ class Base(object):
 
     def test_reindex_base(self):
         idx = self.create_index()
-        expected = np.arange(idx.size)
+        expected = np.arange(idx.size, dtype=np.intp)
 
         actual = idx.get_indexer(idx)
         tm.assert_numpy_array_equal(expected, actual)