… doc/source/whatsnew/v0.19.0.txt
… pandas/tests/indexes/test_range.py
… pandas/tests/test_algos.py
… pandas/tests/test_base.py
… pandas/tests/test_groupby.py
… pandas/tseries/tests/test_period.py
… pandas/tseries/tests/test_timedeltas.py
… pandas/tseries/tests/test_timeseries.py