@jreback 2016-08-12
yes this looks reasonable. go ahead an make ready on windows. Also add an asv as above.