@jreback 2016-08-13
pls update the docs as indicated. otherwise lgtm.