@chris-b1 2016-08-14
What else were you looking for in the docs? I did add an issue ref in the first paragraph.