@chris-b1 2016-08-15
The changes I just pushed were to get this passing on 32 bit windows.  I don't have anything handy to test 32 bit Linux, I can probably set up a VM, although I think it should generally act the same 32 bit Windows (`np.int_ == np.intp == np.int32`)