@jreback 2016-08-15
ok I'll run on 32 bit tomorrow

I just use a macosx / vm